Snip closing HTML tags from the end of an URL.
There is no way to pair them. Options are to end
the URL on "<" character, or to check the URL ending
for </tag>. This still allows URLs such as:

"link":http://example.com/</foobar>/bar

But actual HTML pairing with regular expression is
futile. Truly we could match pairs, but chevrons will
easily be alone in the URL. We could also match
tag like instances, and cut off URL only when there
is no match and otherwise eat the span. The problem
is that this easily leads to broken HTML since
we can not know if there opening tag elsewhere, and
even then we can not compact broken HTML in the
document itself.