Refactor list parsing
Clean up the old code and do matching in two passes to simplify and
speed up the logic. Now it does not pick-up list-like markup instances,
such as strongs, as list items, but joins them to the previous content.

Line break insertion and inline tag parsing is still not optimal and
possibly causes hard-to-debug issues; both should be run on the each
list items content rather than the whole block to avoid the error-prone
context-unaware inline-syntax-like matching exclusion.

I've for now removed the exclusion filtering, as it should not match
anything valid.

Fixes #172